refactor(test): 가독성을 위한 개행 추가
