+++ src/test/java/chess/BoardInitializerTest.java
@@ -6,12 +6,13 @@ import chess.domain.board.File;
 import chess.domain.board.Rank;
 import chess.domain.board.Square;
 import chess.domain.piece.*;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class BoardInitializerTest {
     @ParameterizedTest(name = "초기 기물을 위치에 맞게 생성한다.")
@@ -20,7 +21,7 @@ public class BoardInitializerTest {
         Chessboard chessboard = new Chessboard();
         BoardInitializer.initializeBoard(chessboard);
 
-        Assertions.assertThat(chessboard.getPieceAt(Square.getInstanceOf(file, rank)).getPieceType())
+        assertThat(chessboard.getPieceAt(Square.getInstanceOf(file, rank)).getPieceType())
                 .isEqualTo(piece.getPieceType());
     }
