+++ src/test/java/chess/piece/KingTest.java
@@ -6,9 +6,10 @@ import chess.domain.board.Square;
 import chess.domain.piece.Camp;
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceType;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 class KingTest {
     private final Piece king = PieceType.KING.createPiece(Camp.WHITE);
@@ -19,7 +20,7 @@ class KingTest {
         Square source = Square.getInstanceOf(File.A, Rank.ONE);
         Square target = Square.getInstanceOf(File.A, Rank.THREE);
 
-        Assertions.assertThat(king.canMove(source, target))
+        assertThat(king.canMove(source, target))
                 .isFalse();
     }
 
@@ -29,7 +30,7 @@ class KingTest {
         Square source = Square.getInstanceOf(File.A, Rank.ONE);
         Square target = Square.getInstanceOf(File.A, Rank.TWO);
 
-        Assertions.assertThat(king.canMove(source, target))
+        assertThat(king.canMove(source, target))
                 .isTrue();
     }
 
@@ -39,7 +40,7 @@ class KingTest {
         Square source = Square.getInstanceOf(File.A, Rank.ONE);
         Square target = Square.getInstanceOf(File.B, Rank.ONE);
 
-        Assertions.assertThat(king.canMove(source, target))
+        assertThat(king.canMove(source, target))
                 .isTrue();
     }
 
@@ -49,8 +50,7 @@ class KingTest {
         Square source = Square.getInstanceOf(File.A, Rank.ONE);
         Square target = Square.getInstanceOf(File.B, Rank.TWO);
 
-        Assertions.assertThat(king.canMove(source, target))
+        assertThat(king.canMove(source, target))
                 .isTrue();
     }
-
 }
