+++ src/test/java/chess/piece/PieceTest.java
@@ -1,20 +1,20 @@
-package chess;
+package chess.piece;
 
 import chess.domain.piece.*;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.stream.Stream;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 public class PieceTest {
-    // 리펙토링
+
     @ParameterizedTest(name = "각 체스 기물을 생성할 수 있다.")
     @MethodSource("pieceProvider")
     void createPieceSuccessTest(Piece piece) {
-        Assertions.assertThat(piece)
-                .isNotNull();
+        assertThat(piece).isNotNull();
     }
 
     static Stream<Arguments> pieceProvider() {
